restructure javascript directories and drop mailcheck
