+++ app/assets/javascripts/active_admin.js
@@ -1,0 +1,1 @@
+//= require active_admin/base
