+++ app/assets/javascripts/campaigner-facing/ajax.js
@@ -1,4 +1,4 @@
-let ErrorDisplay = require('show_errors');
+let ErrorDisplay = require('shared/show_errors');
 
 $(function(){
   var handleStart = function(e,i){
