+++ app/assets/javascripts/campaigner-facing/analytics.js
