+++ app/assets/javascripts/campaigner-facing/collection_editor.js
@@ -8,8 +8,8 @@
 //   Twitter's typeahead: http://twitter.github.io/typeahead.js/
 //   - Use for autocompleting for setting the field's name value
 
-const setupOnce = require('setup_once');
-const ErrorDisplay = require('show_errors');
+const setupOnce = require('campaigner-facing/setup_once');
+const ErrorDisplay = require('shared/show_errors');
 
 (function(){
   let CollectionEditor = Backbone.View.extend({
