+++ app/assets/javascripts/campaigner-facing/configure_wysiwyg.js
