+++ app/assets/javascripts/campaigner-facing/dropzone_image_upload.js
