+++ app/assets/javascripts/campaigner-facing/form_preview.js
