+++ app/assets/javascripts/campaigner-facing/search.js
