+++ app/assets/javascripts/campaigner-facing/selectize_config.js
