+++ app/assets/javascripts/campaigner-facing/setup_once.js
