+++ app/assets/javascripts/campaigner-facing/syntax-highlighting.js
