+++ app/assets/javascripts/campaigner-facing/tooltips.js
