--- app/assets/javascripts/liquid.js
@@ -1,8 +1,0 @@
-//= require jquery
-//= require jquery_ujs
-//= require show_errors
-//= require bootstrap-sprockets
-
-// These two files combine to do error suggestions for mistyped emails.
-//= require mailcheck
-//= require email_check
