+++ app/assets/javascripts/member-facing/backbone/action_counter.js
