+++ app/assets/javascripts/member-facing/backbone/braintree_hosted_fields.js
