+++ app/assets/javascripts/member-facing/backbone/campaigner_overlay.js
