+++ app/assets/javascripts/member-facing/backbone/currency_methods.js
