+++ app/assets/javascripts/member-facing/backbone/mobile_check.js
