+++ app/assets/javascripts/member-facing/backbone/overlay_toggle.js
