+++ app/assets/javascripts/member-facing/backbone/simple_carousel.js
