+++ app/assets/javascripts/member-facing/backbone/thermometer.js
