+++ app/assets/javascripts/shared/pub_sub.js
