+++ app/assets/javascripts/shared/show_errors.js
