+++ app/views/layouts/liquid.slim
@@ -6,7 +6,6 @@ html
     meta property='fb:app_id' content="#{Settings.facebook_app_id}"
 
     = stylesheet_link_tag "https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css"
-    = javascript_include_tag "liquid"
     = csrf_meta_tags
 
   body#page-top.index
