+++ app/views/layouts/sumofus.slim
@@ -5,7 +5,7 @@ html lang="#{@page.language.try(:code).try(:downcase)}"
     meta name="viewport" content="width=device-width, initial-scale=1.0"
     meta name="description" content=t('branding.description')
     = stylesheet_link_tag "member-facing"
-    = javascript_include_tag "sumofus"
+    = javascript_include_tag "member-facing"
 
     - unless @page.optimizely_disabled?
       = render partial: 'shared/optimizely_snippet'
