+++ config/initializers/assets.rb
@@ -10,8 +10,4 @@ Rails.application.config.assets.version = '1.0'
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
 # Rails.application.config.assets.precompile += %w( search.js )
-Rails.application.config.assets.precompile += %w(
-  liquid.js
-  member-facing.css
-  sumofus.js
-)
+Rails.application.config.assets.precompile += %w( member-facing.css member-facing.js )
